Testing: Move testing-library helper
