+++ client/components/screen-options-tab/test/index.js
@@ -5,7 +5,7 @@
 import { fireEvent, screen } from '@testing-library/react';
 import jetpack from 'calypso/state/jetpack/reducer';
 import { reducer as ui } from 'calypso/state/ui/reducer';
-import { render as rtlRender } from 'calypso/test-helpers/config/testing-library';
+import { render as rtlRender } from 'calypso/test-helpers/testing-library';
 import ScreenOptionsTab from '../index';
 
 jest.mock( 'react-redux', () => ( {
