+++ client/jetpack-cloud/sections/settings/loading/test/index.jsx
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 
-import { render } from 'calypso/test-helpers/config/testing-library';
+import { render } from 'calypso/test-helpers/testing-library';
 import '@testing-library/jest-dom/extend-expect';
 import AdvancedCredentialsLoadingPlaceholder from '../index';
